adds JWT claims

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 
@@ -17,7 +17,7 @@ class Item(Resource):
                         help = "Every item needs a store ID!"
                         )
 
-    @jwt_required
+    @jwt_required()
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -41,7 +41,12 @@ class Item(Resource):
 
         return item.json(), 201
     
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin previlage required.'}, 402
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
